agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph==0.0.26
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.6
langchain-community==0.0.17
tavily-python==0.3.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
tiktoken==0.5.2
pinecone-client==3.0.2
pymongo==4.6.1
pgvector==0.2.4
numpy==1.26.3
